Conflicts:
	CCTM/src/util/util/setup_logdev.F

	modified:   CCTM/docs/README.md
	new file:   CCTM/docs/Release_Notes/Aerosol_ICBC_GetPar.md
	new file:   CCTM/docs/Release_Notes/Bioseason.md
	new file:   CCTM/docs/Release_Notes/Combine_Math.md
	new file:   CCTM/docs/Release_Notes/Compilation_Options_and_Debug.md
	new file:   CCTM/docs/Release_Notes/Duplicate_N2O5.md
	new file:   CCTM/docs/Release_Notes/Emiss_Check_Zero_PT_Sources.md
	new file:   CCTM/docs/Release_Notes/Grid_Check_Tolerance.md
	new file:   CCTM/docs/Release_Notes/HDIFF_hang.md
	new file:   CCTM/docs/Release_Notes/Hay_Land_Use.md
	new file:   CCTM/docs/Release_Notes/INTERPX.md
	new file:   CCTM/docs/Release_Notes/Intel17_Solver_Compatibility.md
	modified:   CCTM/docs/Release_Notes/README.md
	new file:   CCTM/docs/Release_Notes/SWAP_routines.md
	new file:   CCTM/docs/Release_Notes/SpecDefExtraction.md
	new file:   CCTM/docs/Release_Notes/Two_Way_Convective_Scheme.md
	modified:   CCTM/scripts/bldit_cctm.csh
	modified:   CCTM/scripts/run_cctm.csh
	modified:   CCTM/src/phot/inline/CLOUD_OPTICS.F
	modified:   CCTM/src/util/util/setup_logdev.F
	new file:   DOCS/Developers_Guide/CMAQ_Dev_Guide.md
	new file:   DOCS/Developers_Guide/PDF/CMAQ_Dev_Guide.pdf
	deleted:    DOCS/Known_Issues/CMAQv5.2-i2/LTNG_DEFN.F
	deleted:    DOCS/Known_Issues/CMAQv5.2-i4/advstep.F
	deleted:    DOCS/Known_Issues/CMAQv5.2-i5/AEROSOL_CHEMISTRY.F
	deleted:    DOCS/Known_Issues/CMAQv5.2-i5/AERO_DATA.F
	deleted:    DOCS/Known_Issues/CMAQv5.2-i5/SOA_DEFN.F
	deleted:    DOCS/Known_Issues/CMAQv5.2-i5/aero_driver.F
	deleted:    DOCS/Known_Issues/CMAQv5.2-i5/aero_subs.F
	deleted:    DOCS/Known_Issues/CMAQv5.2-i5/getpar.f
	modified:   DOCS/Known_Issues/README.md
	modified:   DOCS/README.md
	new file:   DOCS/Tutorials/CMAQ_DustInput.md
	modified:   DOCS/Tutorials/README.md
	new file:   DOCS/User_Manual/Appendix_A/cb05e51_ae6_aq/AE6_species_table.md
	modified:   DOCS/User_Manual/Appendix_A/cb05e51_ae6_aq/cb05e51_ae6_aq_species_table.md
	new file:   DOCS/User_Manual/Appendix_A/cb05tucl/AE6_species_table.md
	modified:   DOCS/User_Manual/Appendix_A/cb05tucl/cb05tucl_species_table.md
	new file:   DOCS/User_Manual/Appendix_A/cb6mp_ae6_aq/AE6_species_table.md
	new file:   DOCS/User_Manual/Appendix_A/cb6mp_ae6_aq/CB6_species_table.md
	new file:   DOCS/User_Manual/Appendix_A/cb6r3_ae6_aq/AE6_species_table.md
	modified:   DOCS/User_Manual/Appendix_A/cb6r3_ae6_aq/CB6_species_table.md
	new file:   DOCS/User_Manual/Appendix_A/racm2_ae6_aq/AE6_species_table.md
	modified:   DOCS/User_Manual/Appendix_A/racm2_ae6_aq/racm2_ae6_aq_species_table.md
	new file:   DOCS/User_Manual/Appendix_A/racm2_ae6_aq/racm2_ae6_aq_species_table_aerosol.md
	new file:   DOCS/User_Manual/Appendix_A/saprc07tb_ae6_aq/AE6_species_table.md
	modified:   DOCS/User_Manual/Appendix_A/saprc07tb_ae6_aq/saprc07tb_ae6_aq_species_table.md
	new file:   DOCS/User_Manual/Appendix_A/saprc07tb_ae6_aq/saprc07tb_ae6_aq_species_table_aerosol.md
	new file:   DOCS/User_Manual/Appendix_A/saprc07tc_ae6_aq/AE6_species_table.md
	modified:   DOCS/User_Manual/Appendix_A/saprc07tc_ae6_aq/saprc07tc_ae6_aq_species_table.md
	new file:   DOCS/User_Manual/Appendix_A/saprc07tc_ae6_aq/saprc07tc_ae6_aq_species_table_aerosol.md
	new file:   DOCS/User_Manual/Appendix_A/saprc07tic_ae6i_aq/AE6_species_table.md
	modified:   DOCS/User_Manual/Appendix_A/saprc07tic_ae6i_aq/saprc07tic_ae6i_aq_species_table.md
	new file:   DOCS/User_Manual/Appendix_A/saprc07tic_ae6i_aq/saprc07tic_ae6i_aq_species_table_aerosol.md
	modified:   DOCS/User_Manual/CMAQ_OGD_appendix_A.md
	modified:   DOCS/User_Manual/CMAQ_OGD_ch01_intro.md
	modified:   DOCS/User_Manual/CMAQ_OGD_ch02_overview.md
	modified:   DOCS/User_Manual/CMAQ_OGD_ch03_features.md
	modified:   DOCS/User_Manual/CMAQ_OGD_ch04_science.md
	modified:   DOCS/User_Manual/CMAQ_OGD_ch05_sys_req.md
	modified:   DOCS/User_Manual/CMAQ_OGD_ch07_programs_libraries.md
	modified:   DOCS/User_Manual/CMAQ_OGD_ch09_grid_defn.md
	modified:   DOCS/User_Manual/CMAQ_OGD_ch10_new_simulation.md
	modified:   DOCS/User_Manual/CMAQ_OGD_ch11_code_management.md
	modified:   DOCS/User_Manual/CMAQ_OGD_ch12_analysis_tools.md
	modified:   DOCS/User_Manual/CMAQ_OGD_quick_start.md
	modified:   DOCS/User_Manual/README.md
	modified:   POST/appendwrf/scripts/bldit_appendwrf.csh
	modified:   POST/appendwrf/scripts/run_appendwrf.csh
	modified:   POST/bldoverlay/scripts/bldit_bldoverlay.csh
	modified:   POST/bldoverlay/scripts/run_bldoverlay.csh
	modified:   POST/block_extract/scripts/bldit_block_extract.csh
	modified:   POST/block_extract/scripts/run_block_extract.csh
	modified:   POST/combine/scripts/bldit_combine.csh
	modified:   POST/combine/scripts/run_combine.csh
	modified:   POST/combine/scripts/spec_def_files/linkem
	modified:   POST/hr2day/scripts/bldit_hr2day.csh
	modified:   POST/hr2day/scripts/run_hr2day.csh
	modified:   POST/sitecmp/scripts/bldit_sitecmp.csh
	modified:   POST/sitecmp/scripts/run_sitecmp_AQS_Hourly.csh
	modified:   POST/sitecmp_dailyo3/scripts/bldit_sitecmp_dailyo3.csh
	modified:   POST/sitecmp_dailyo3/scripts/run_sitecmp_dailyo3_AQS.csh
	modified:   POST/sitecmp_dailyo3/scripts/run_sitecmp_dailyo3_CASTNET.csh
	modified:   POST/writesite/scripts/bldit_writesite.csh
	modified:   POST/writesite/scripts/run_writesite.csh
	modified:   PREP/mcip/docs/CHANGES
	modified:   PREP/mcip/docs/ReleaseNotes
	modified:   PREP/mcip/src/Makefile
	modified:   PREP/mcip/src/alloc_ctm.f90
	modified:   PREP/mcip/src/alloc_met.f90
	modified:   PREP/mcip/src/alloc_x.f90
	modified:   PREP/mcip/src/blddesc.f90
	modified:   PREP/mcip/src/comheader.f90
	new file:   PREP/mcip/src/comheader_lufrac.f90
	new file:   PREP/mcip/src/comheader_mos.f90
	new file:   PREP/mcip/src/comheader_soi.f90
	modified:   PREP/mcip/src/dealloc_ctm.f90
	modified:   PREP/mcip/src/dealloc_met.f90
	modified:   PREP/mcip/src/dealloc_x.f90
	modified:   PREP/mcip/src/files_mod.f90
	modified:   PREP/mcip/src/getluse.f90
	new file:   PREP/mcip/src/getpsih.f90
	modified:   PREP/mcip/src/init_met.f90
	modified:   PREP/mcip/src/init_x.f90
	new file:   PREP/mcip/src/lucro.f90
	new file:   PREP/mcip/src/luoutcom_mod.f90
	modified:   PREP/mcip/src/luvars_mod.f90
	modified:   PREP/mcip/src/mcip.f90
	modified:   PREP/mcip/src/mcipparm_mod.f90
	modified:   PREP/mcip/src/mcoutcom_mod.f90
	modified:   PREP/mcip/src/metcro.f90
	modified:   PREP/mcip/src/metvars2ctm.f90
	modified:   PREP/mcip/src/metvars_mod.f90
	new file:   PREP/mcip/src/moscro.f90
	new file:   PREP/mcip/src/mosoutcom_mod.f90
	modified:   PREP/mcip/src/pblsup.f90
	modified:   PREP/mcip/src/rdwrfem.f90
	new file:   PREP/mcip/src/run_mcip.csh
	modified:   PREP/mcip/src/sat2mcip_mod.f90
	modified:   PREP/mcip/src/setgriddefs.f90
	modified:   PREP/mcip/src/setup_mm5v3.f90
	modified:   PREP/mcip/src/setup_wrfem.f90
	new file:   PREP/mcip/src/soilcro.f90
	new file:   PREP/mcip/src/soioutcom_mod.f90
	modified:   PREP/mcip/src/wrf_netcdf_mod.f90
	modified:   PREP/mcip/src/wrfemopts.f90
	modified:   PREP/mcip/src/xvars_mod.f90
	modified:   README.md
	modified:   UTIL/create_ebi/README.md
	deleted:    UTIL/create_ebi/scripts/bldit.create_ebi.csh
	new file:   UTIL/create_ebi/scripts/bldrun_create_ebi.csh
	deleted:    UTIL/create_ebi/scripts/run.create_ebi.csh
	modified:   UTIL/inline_phot_preproc/README.md
	deleted:    UTIL/inline_phot_preproc/scripts/bldit.inline_phot_preproc.csh
	renamed:    UTIL/inline_phot_preproc/scripts/run.inline_phot_preproc.csh -> UTIL/inline_phot_preproc/scripts/bldrun.inline_phot_preproc.csh
	modified:   bldit_project.csh
	modified:   config_cmaq.csh

+++ UTIL/inline_phot_preproc/scripts/bldrun.inline_phot_preproc.csh
@@ -1,23 +1,23 @@
 #!/bin/csh -fx
 
-source ../../../config.cmaq
+ set echo
 
-#> CMAQv5.1 Mechanism Options: cb05tucl_ae6_aq cb05tump_ae6_aq cb05e51_ae6_aq saprc07tb_ae6_aq saprc07tc_ae6_aq  saprc07tic_ae6i_aq   racm2_ae6_aq 
- set Mechanism = racm2_ae6_aq          #> CMAQ mechanism ID
- set REPO  = ../../CCTM/src/MECHS      #> Repository Directory with all mechanism data modules
- set BASE  = $CMAQ_HOME/UTIL/inline_phot_preproc
- set XBASE = $BASE/scripts/BLD_$Mechanism
+#> CMAQ Mechanism under Repository directory CTM/src/MECHS or
+#> user defines their own Mechanism name
+ set Mechanism = cb6r3_ae6_aq          #> CMAQ mechanism ID
+# path to CMAQ repository  
+ set REPO  = ../../CCTM/src/MECHS
+ set BASE  = $cwd
+ set XBASE = $BASE
 
- set CSQY_DIR    = ${BASE}/photolysis_CSQY_data
- set REFRACT_DIR = ${BASE}/water_clouds
- set WVBIN_DIR   = ${BASE}/flux_data
-
+#> option to set compiler and build a new executable (not required)
+ setenv COMPILER  GFORT  #> INTEL, PGF90, or GFORT
  set day = ` date "+%b-%d-%Y" `
- set OUTDIR = $BASE/output/csqy_table_${Mechanism}-${day}
+ set OUTDIR = $BASE/output/csqy_table_${Mechanism}-${day}-${COMPILER}
 
 #> specify directory containing the mechanism modules or include files
- setenv GC_INC $REPO/$Mechanism         #> Repository Directory with the above mechanism data modules 
-#setenv GC_INC $BASE/input/$Mechanism   #> User Defined Directory with the above mechanism data modules
+  setenv GC_INC $BASE/input/$Mechanism #> User Defines directory with the below mechanism data module
+# setenv GC_INC $REPO/$Mechanism
 
 #> use RXNS_DATA_MODULE, comment out if CMAQ v5.02 and keep if CMAQ v5.1 or higher
  setenv USE_RXNS_MODULES T
@@ -36,9 +36,19 @@ source ../../../config.cmaq
 #>  set F if CMAQ v5.02 and T if CMAQ v5.1 or higher
  setenv SPLIT_OUTPUT T
 
+#> define exectubale
+ set EXEC = CSQY_TABLE_PROCESSOR_${Mechanism}
+
+#> create executable
+ setenv APPL $Mechanism
+ cd src ; make clean; make -f dumb.makefile; cd ../
+
  if( ! ( -d $OUTDIR ) ) mkdir -p $OUTDIR
 
- cd $BASE
+ set CSQY_DIR    = ${BASE}/photolysis_CSQY_data
+ set REFRACT_DIR = ${BASE}/water_clouds
+ set WVBIN_DIR   = ${BASE}/flux_data
+
  ln -s -f $CSQY_DIR  CSQY_DATA_RAW
  ln -s -f $WVBIN_DIR/wavel-bins.dat  WVBIN_FILE
  ln -s -f $WVBIN_DIR/solar-p05nm-UCI.dat  FLUX_FILE
@@ -52,8 +62,18 @@ source ../../../config.cmaq
 #ln -s -f $REFRACT_DIR/soot00 SOOT
  ln -s -f $REFRACT_DIR/soot00-two_way-Oct_21_2012 SOOT
  ln -s -f $REFRACT_DIR/ssam00 SEASALT
- 
- $XBASE/CSQY_TABLE_PROCESSOR
+
+# cd $OUTDIR
+
+ if( ! ( -e  $XBASE/$EXEC ) )then
+     \ls $XBASE/$EXEC
+     echo "make failed or value of XBASE incorrect"
+     exit()
+ endif
+
+
+ $XBASE/$EXEC >&! bldrun.log
+
 
  \rm -f CSQY_DATA_RAW
  \rm -f WVBIN_FILE
@@ -66,6 +86,18 @@ source ../../../config.cmaq
  \rm -f SOOT
  \rm -f SEASALT
  \rm -f DUST
+ \rm -f $XBASE/$EXEC
  \rm -f fort.*
+#\rm -f OPTICS_DATA 
+#\rm -f CSQY_DATA 
  
+ cd $BASE
+
+unset echo
+echo " "
+echo " "
+echo "Check directory ${OUTDIR} for CSQY_DATA_${Mechanism} and PHOT_OPTICS.dat files"
+echo " "
+echo " "
+
  exit()
